Conflicts:
	autoconf/configure.ac
	configure
	include/llvm/ADT/Triple.h
	include/llvm/Support/ELF.h
	lib/Support/Triple.cpp
	projects/sample/autoconf/configure.ac
	projects/sample/configure

+++ lldb/source/Commands/CommandObjectRegister.cpp
@@ -29,6 +29,7 @@
 #include "lldb/Target/ExecutionContext.h"
 #include "lldb/Target/Process.h"
 #include "lldb/Target/RegisterContext.h"
+#include "lldb/Target/SectionLoadList.h"
 #include "lldb/Target/Thread.h"
 
 using namespace lldb;
@@ -201,7 +202,7 @@ protected:
                     }
                     else
                     {
-                        result.AppendErrorWithFormat ("invalid register set index: %zu\n", set_idx);
+                        result.AppendErrorWithFormat("invalid register set index: %" PRIu64 "\n", (uint64_t)set_idx);
                         result.SetStatus (eReturnStatusFailed);
                         break;
                     }
